Fixed the Swiss system
it should pair the team based on performance and not ELO which i overlooked. Added loggers to check and this code works. I need to reimplement the DoubleElimination as well as generate a method for the swiss system? make tournament_team entity to store intermediary database and seperate all the enums in the enum folder